% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{antirect_resblock_2d}
\alias{antirect_resblock_2d}
\title{antirectifier resblock with bottleneck architecture}
\usage{
antirect_resblock_2d(
  x,
  filters = 32L,
  kernel_size = c(3L, 3L),
  downsample = FALSE
)
}
\description{
antirectifier resblock with bottleneck architecture
}
