% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencoder_model}
\alias{autoencoder_model}
\title{Autoencoder keras model wrapper over simple dense layers}
\usage{
autoencoder_model(
  original_dim,
  hidden_dims,
  num_layers = 3,
  regularizer = NULL,
  use_batchnorm = FALSE,
  name = NULL
)
}
\arguments{
\item{original_dim}{integer denoting original dimensionality of input space}

\item{hidden_dims}{may be an integer, or a list containing desired units for
encoder step.  `list(512, 256, 64)` will give a `code_dimension` of `64`}

\item{num_layers}{integer, number of layers to use in each encoder decoder.
[num_layers] must equal [length(hidden_dims)] unless hidden_dims is a scalar,
in which case, hidden_dims is automatically computed based on nearest powers of 2}

\item{regularizer}{callable or string name of keras regularizer}

\item{use_batchnorm}{bool indicating use of batchnorm layer inbetween each dense}

\item{name}{(optional) string to name the model}
}
\description{
Autoencoder keras model wrapper over simple dense layers
}
