% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_and_compile}
\alias{build_and_compile}
\title{Convencience fun to one-liner a model with in-outs}
\usage{
build_and_compile(
  input,
  output,
  optimizer = "adam",
  loss = "mse",
  metric = "acc"
)
}
\description{
Convencience fun to one-liner a model with in-outs
}
