% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{layer_activation_belu}
\alias{layer_activation_belu}
\title{keras layer-lambda Bipolar ELU as in https://arxiv.org/abs/1709.04054.}
\usage{
layer_activation_belu(object, name = "belu")
}
\description{
keras layer-lambda Bipolar ELU as in https://arxiv.org/abs/1709.04054.
}
