% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{layer_activation_brelu}
\alias{layer_activation_brelu}
\title{Bipolar ReLU as in https://arxiv.org/abs/1709.04054}
\usage{
layer_activation_brelu(object, name = "brelu")
}
\description{
Bipolar ReLU as in https://arxiv.org/abs/1709.04054
}
