% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{layer_activation_nac}
\alias{layer_activation_nac}
\title{keras lambda layer implementation of NAC as in https://arxiv.org/abs/1808.00508}
\usage{
layer_activation_nac(object, depth, name = "nac")
}
\description{
keras lambda layer implementation of NAC as in https://arxiv.org/abs/1808.00508
}
