% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{layer_activation_nalu}
\alias{layer_activation_nalu}
\title{keras lambda-layer NALU as in https://arxiv.org/abs/1808.00508}
\usage{
layer_activation_nalu(
  object,
  depth,
  epsilon = 1e-30,
  name = NULL,
  reuse = NULL
)
}
\description{
keras lambda-layer NALU as in https://arxiv.org/abs/1808.00508
}
