% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antirectifier.R
\name{layer_antirectifier}
\alias{layer_antirectifier}
\title{Antirectifier layer wrapper for keras (thanks fchollet).}
\usage{
layer_antirectifier(object)
}
\description{
Expects input shape of (examples, samples), normalizing over the `samples` dimension.
}
