% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antirectifier.R
\name{layer_antirectifier_nd}
\alias{layer_antirectifier_nd}
\title{Antirectifier layer wrapper for keras (thanks fchollet) in arbitrary dimensions.}
\usage{
layer_antirectifier_nd(object, axis = 3L)
}
\arguments{
\item{axis}{integer. The `samples` dimension to normalize over, defaults to 3 for 2d input.
NOTE: Axis is `R` style 1-based indexing, e.g. axis = 2 refers to the 2nd dim, not 3rd.}
}
\description{
Antirectifier layer wrapper for keras (thanks fchollet) in arbitrary dimensions.
}
