% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.R
\name{layer_causal_conv1d}
\alias{layer_causal_conv1d}
\title{Causal convolution layer, masks out future (look-ahead) sequences}
\usage{
layer_causal_conv1d(
  object,
  filters,
  kernel_size,
  strides = 1L,
  dilation_rate = 1L,
  activation = NULL,
  use_bias = TRUE,
  kernel_initializer = tf$keras$initializers$glorot_uniform(),
  bias_initializer = tf$keras$initializers$zeros(),
  kernel_regularizer = NULL,
  bias_regularizer = NULL,
  activity_regularizer = NULL,
  kernel_constraint = NULL,
  bias_constraint = NULL,
  trainable = TRUE,
  name = "causal_conv1d",
  ...
)
}
\description{
Causal convolution layer, masks out future (look-ahead) sequences
}
