% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glu.R
\name{layer_glu}
\alias{layer_glu}
\title{Gated-Linear-Unit keras lambda layer}
\usage{
layer_glu(
  object,
  filters,
  kernel_size,
  kernel_initializer = "glorot_normal",
  kernel_regularizer = NULL,
  bias_initializer = "zeros",
  bias_regularizer = NULL,
  name = NULL,
  trainable = TRUE
)
}
\description{
Computes gated temporal convolutions as in: `hl(X) = (X∗W+b) ⊗ σ(X∗V+c)``,
where `X` is a data matrix, `W` is a weight kernel, and `b`,`c` are biases.
`⊗` is the element-wise multiplication operator
}
\references{
https://arxiv.org/abs/1612.08083
}
