% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glu.R
\name{layer_glu_block}
\alias{layer_glu_block}
\title{Gated-Linear-Unit block}
\usage{
layer_glu_block(
  object,
  filters,
  num_layers = 3,
  kernel_size = 3,
  kernel_initializer = "glorot_normal",
  kernel_regularizer = NULL,
  bias_initializer = "zeros",
  bias_regularizer = NULL,
  name = NULL,
  trainable = TRUE
)
}
\description{
Builds upon `layer_glu()`, combining `n` layers into a single block with
residual connections
}
\references{
https://arxiv.org/abs/1612.08083
}
