% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pru.R
\name{layer_pyramidal_recurrent_block}
\alias{layer_pyramidal_recurrent_block}
\title{Pyramidal recurrent block}
\usage{
layer_pyramidal_recurrent_block(
  object,
  units,
  num_layers = 3,
  cell_type = "LSTM",
  activation = "tanh",
  projection_activation = TRUE,
  projection_batchnorm = TRUE,
  kernel_initializer = "glorot_normal",
  kernel_regularizer = NULL,
  recurrent_activation = "sigmoid",
  recurrent_initializer = "orthogonal",
  recurrent_regularizer = NULL,
  bias_initializer = "zeros",
  bias_regularizer = NULL,
  recurrent_dropout = 0,
  name = NULL,
  trainable = TRUE
)
}
\arguments{
\item{units}{Number of hidden units for each layer in the block}

\item{cell_type}{string denoting type of cell to use. default: `LSTM`}

\item{activation}{activation function for each layer}
}
\description{
Builds a block of `num_layers` pyramidal recurrent layers, which
}
\references{
https://arxiv.org/pdf/1808.09029.pdf
}
