% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{resblock_1d}
\alias{resblock_1d}
\title{Specialized Residual unit which contains a linear projection layer up-front with
two conv blocks with residual connections between before a maxpool}
\usage{
resblock_1d(x, filters, kernel_size)
}
\description{
Specialized Residual unit which contains a linear projection layer up-front with
two conv blocks with residual connections between before a maxpool
}
\references{
https://arxiv.org/abs/1906.04459
}
