% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{resblock_2d}
\alias{resblock_2d}
\title{Specialized Residual unit which contains a linear projection layer up-front with
two conv blocks with residual connections between before a maxpool}
\usage{
resblock_2d(x, filters, kernel_size = c(3L, 3L), strides = c(2L, 1L))
}
\description{
Specialized Residual unit which contains a linear projection layer up-front with
two conv blocks with residual connections between before a maxpool
}
\references{
https://arxiv.org/abs/1906.04459
}
