% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{resblock_batchnorm_1d}
\alias{resblock_batchnorm_1d}
\title{batchnorm version of `resblock_1d()`}
\usage{
resblock_batchnorm_1d(x, filters, kernel_size = 3L, downsample = TRUE)
}
\description{
batchnorm version of `resblock_1d()`
}
\references{
https://arxiv.org/abs/1906.04459
}
