% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{resblock_batchnorm_1d_base}
\alias{resblock_batchnorm_1d_base}
\title{base unit resblock with batchnorm from Exploring Normalization FB paper (01/2017)}
\usage{
resblock_batchnorm_1d_base(x, filters, kernel_size = 3L)
}
\description{
base unit resblock with batchnorm from Exploring Normalization FB paper (01/2017)
}
