% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{resblock_batchnorm_1d_v2}
\alias{resblock_batchnorm_1d_v2}
\title{As suggsted in original batchnorm paper; computes `g(BN(Wx + b))`}
\usage{
resblock_batchnorm_1d_v2(x, filters, kernel_size = 3L, downsample = TRUE)
}
\description{
As suggsted in original batchnorm paper; computes `g(BN(Wx + b))`
}
\references{
https://arxiv.org/abs/1502.03167
}
