% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{resblock_batchnorm_bottle_1d}
\alias{resblock_batchnorm_bottle_1d}
\title{resblock with bottleneck from the Exploring Normalization FB paper (01/2017)}
\usage{
resblock_batchnorm_bottle_1d(
  x,
  filters = 64L,
  kernel_size = 3L,
  downsample = TRUE
)
}
\description{
resblock with bottleneck from the Exploring Normalization FB paper (01/2017)
}
