% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{resblock_bottle_vanilla}
\alias{resblock_bottle_vanilla}
\title{Resnet block bottleneck from the original paper (1512.03385)}
\usage{
resblock_bottle_vanilla(input, blocks = 1L, filters = c(32, 128))
}
\description{
Resnet block bottleneck from the original paper (1512.03385)
}
\references{
https://arxiv.org/abs/1512.03385
}
